requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# AI
openai==1.55.0

# Image processing (OCR) - optional, install separately if needed
# pillow==10.4.0

# Data export
openpyxl==3.1.5
python-multipart==0.0.9

# Validation
pydantic==2.9.2

# Templates
jinja2==3.1.4

# Testing
pytest==8.3.3
httpx==0.27.2

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
